New runner state: Live
A runner is live if it is running but not yet
promotable to master. One example would be a
service that needs to synchronize a database
before being able to run as master, but can
be considered "live" for other purposes.

Now, services need to complete one successful
status check before being considered promotable.

If the status check returns the Live error code,
the service will continue to stay in the unpromotable
state.