BUG: Don't stop newly spawned action right away
While checking exit codes a new action can be spawned if there is
a mismatch between expected state and real state. This newly spawned
action was incorrectly stopped right away which caused the action
not to block following status checks. So koi was spawning many
instances of the same action waiting in defunct state. Koi is fixed
such that a newly spawned action will be stopped only if the action
script has completed.